Parse database files
Once we've fetched the database files from /empeg/var, we need to start
parsing them into something that we can work with.

This commit parses the 'tags' and 'database' files and converts them
into something useful.

It makes a start on parsing the 'playlists' file. It doesn't convert it
into anything useful, but it does display them to the screen, which is
a start.